+++ packages/HuYaRecorder/src/huya_mobile_api.ts
@@ -1,20 +1,12 @@
 // import { createHash, randomInt } from "node:crypto";
 // import { URLSearchParams } from "node:url";
 
-import axios from "axios";
+import { requester } from "./requester.js";
 
 import { utils } from "@bililive-tools/manager";
 import { assert, getFormatSources } from "./utils.js";
 
 import type { StreamResult, StreamProfile } from "./types.js";
-
-const requester = axios.create({
-  timeout: 10e3,
-  headers: {
-    "User-Agent":
-      "Mozilla/5.0 (Linux; Android 5.0; SM-G900P Build/LRX21T) AppleWebKit/537.36 (KHTML, like Gecko)",
-  },
-});
 
 export async function getRoomInfo(
   roomIdOrShortId: string,
@@ -22,6 +14,12 @@ export async function getRoomInfo(
 ) {
   const res = await requester.get(
     `https://mp.huya.com/cache.php?m=Live&do=profileRoom&roomid=${roomIdOrShortId}`,
+    {
+      headers: {
+        "User-Agent":
+          "Mozilla/5.0 (Linux; Android 5.0; SM-G900P Build/LRX21T) AppleWebKit/537.36 (KHTML, like Gecko)",
+      },
+    },
   );
   const html = res.data;