+++ packages/liveManager/src/recorder.ts
@@ -4,6 +4,8 @@ import { RecorderProvider } from "./manager.js";
 import { AnyObject, PickRequired, UnknownObject } from "./utils.js";
 import { Cache } from "./cache.js";
 
+import type { RecorderType } from "./recorder/index.js";
+
 type FormatName = "auto" | "flv" | "hls" | "fmp4" | "flv_only" | "hls_only" | "fmp4_only";
 type CodecName = "auto" | "avc" | "hevc" | "avc_only" | "hevc_only";
 
@@ -58,7 +60,7 @@ export interface RecorderCreateOpts<E extends AnyObject = UnknownObject> {
   /** 标题关键词，如果直播间标题包含这些关键词，则不会自动录制（仅对斗鱼有效），多个关键词用英文逗号分隔 */
   titleKeywords?: string;
   /** 用于指定录制文件格式，auto时，分段使用ts，不分段使用mp4 */
-  videoFormat?: "auto" | "ts" | "mkv";
+  videoFormat?: "auto" | "ts" | "mkv" | "flv";
   /** 录制类型 */
   recorderType?: "auto" | "ffmpeg" | "mesio" | "bililive";
   /** 流格式优先级 */
@@ -113,6 +115,7 @@ export interface RecordHandle {
   id: string;
   stream: string;
   source: string;
+  recorderType?: RecorderType;
   url: string;
   ffmpegArgs?: string[];
   progress?: Progress;
@@ -128,7 +131,13 @@ export interface DebugLog {
   text: string;
 }
 
-export type GetSavePath = (data: { owner: string; title: string; startTime?: number }) => string;
+export type GetSavePath = (data: {
+  owner: string;
+  title: string;
+  startTime: number;
+  liveStartTime: Date;
+  recordStartTime: Date;
+}) => string;
 
 export interface Recorder<E extends AnyObject = UnknownObject>
   extends Emitter<{
@@ -164,7 +173,7 @@ export interface Recorder<E extends AnyObject = UnknownObject>
     living: boolean;
     owner: string;
     title: string;
-    startTime?: Date;
+    startTime: Date;
     avatar: string;
     cover: string;
     liveId?: string;