+++ packages/liveManager/src/recorder/mesioRecorder.ts
@@ -163,8 +163,8 @@ export class MesioRecorder extends EventEmitter implements IRecorder {
 
     this.command = this.createCommand();
 
-    this.streamManager.on("videoFileCreated", ({ filename, cover, rawFilename }) => {
-      this.emit("videoFileCreated", { filename, cover, rawFilename });
+    this.streamManager.on("videoFileCreated", ({ filename, cover, rawFilename, title }) => {
+      this.emit("videoFileCreated", { filename, cover, rawFilename, title });
     });
     this.streamManager.on("videoFileCompleted", ({ filename }) => {
       this.emit("videoFileCompleted", { filename });
@@ -180,6 +180,7 @@ export class MesioRecorder extends EventEmitter implements IRecorder {
       "--fix",
       "-H",
       "User-Agent: Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/73.0.3683.86 Safari/537.36",
+      "--no-proxy",
     ];
     if (this.debugLevel === "verbose") {
       inputOptions.push("-v");