+++ packages/HuYaRecorder/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.9.0
+
+- `recordHandle` 新增参数 `recorderType`
+- 录制：优化ffmpeg默认参数，fmp4使用m4s后缀 [#224](https://github.com/renmu123/biliLive-tools/pull/224)
+
 # 1.8.0
 
 - 新增`debugLevel`参数，支持`none`、`basic`、`verbose`